+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start:dev": "npx nodemon src/index.ts",
     "start:prod": "webpack --mode production && node ./dist/bundle.js",
-    "start:multi": "",
+    "start:multi": "npx nodemon src/index-bal.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "spiderVS",
